requirements.txt
torch~=2.7.0+cu118
torchvision~=0.22.0+cu118
pandas~=2.2.3
pillow~=11.0.0
tqdm~=4.67.1
nltk~=3.9.1
numpy~=2.1.2
data~=0.4
matplotlib~=3.10.0
pyyaml~=6.0.2